+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ### Unreleased
 
+- feat: fix alias + smtp_forward w/multi-domain
+- test: runner is now node:test
+- test: remove unnecessary done callbacks #20
+
 ### [1.0.3] - 2025-01-30
 
 - prettier: move config into package.json